requirements.txt
python-dotenv
PySide6
spleeter
pycodestyle
